+++ src/onegov.agency/onegov/agency/forms/__init__.py
@@ -1,4 +1,5 @@
 from onegov.agency.forms.agency import ExtendedAgencyForm
+from onegov.agency.forms.agency import MoveAgencyForm
 from onegov.agency.forms.membership import MembershipForm
 from onegov.agency.forms.mutation import MutationForm
 
@@ -7,5 +8,6 @@ __all__ = (
     'ExtendedAgencyForm',
     'ExtendedPersonForm',
     'MembershipForm',
+    'MoveAgencyForm',
     'MutationForm',
 )